rename module to serialization.py
