+++ stix2/datastore/filesystem.py
@@ -9,12 +9,13 @@ import stat
 import six
 
 from stix2 import v20, v21
-from stix2.base import _STIXBase, serialize
+from stix2.base import _STIXBase
 from stix2.datastore import (
     DataSink, DataSource, DataSourceError, DataStoreMixin,
 )
 from stix2.datastore.filters import Filter, FilterSet, apply_common_filters
 from stix2.parsing import parse
+from stix2.serialization import serialize
 from stix2.utils import format_datetime, get_type_from_id, parse_into_datetime
