+++ stix2/serialization.py
